Add custom TH3D Studio NEMA 17 pancake stepper motor
This commit adds the custom stepper motor built for TH3D.
The original manufacturer is unfortunately unknown.
No markings are present on the motor and TH3D is secret about their
supplier/manufacturer.
Given the popularity of TH3D it makes nevertheless sense to add the
motor, as many extruders sold by TH3D ship with this motor.